expand some posible solutions
